+++ delta-coverage-core/src/main/kotlin/io/github/surpsg/deltacoverage/lib/kotlin/ExtraStandard.kt
@@ -1,0 +1,7 @@
+package io.github.surpsg.deltacoverage.lib.kotlin
+
+internal inline fun <T> T.applyIf(condition: Boolean, block: T.() -> T): T = if (condition) {
+    block()
+} else {
+    this
+}