+++ delta-coverage-core/src/main/kotlin/io/github/surpsg/deltacoverage/report/EnabledReportTypeFactory.kt
@@ -1,0 +1,21 @@
+package io.github.surpsg.deltacoverage.report
+
+import io.github.surpsg.deltacoverage.config.ReportsConfig
+
+object EnabledReportTypeFactory {
+
+    fun obtain(reportsConfig: ReportsConfig): Set<ReportType> =
+        ReportType.entries.asSequence()
+            .map { reportType ->
+                when (reportType) {
+                    ReportType.HTML -> ReportType.HTML to reportsConfig.html.enabled
+                    ReportType.XML -> ReportType.XML to reportsConfig.xml.enabled
+                    ReportType.CONSOLE -> ReportType.CONSOLE to reportsConfig.console.enabled
+                    ReportType.MARKDOWN -> ReportType.MARKDOWN to reportsConfig.markdown.enabled
+                    ReportType.CSV -> ReportType.CSV to reportsConfig.csv.enabled
+                }
+            }
+            .filter { (_, enabled) -> enabled }
+            .map { (reportType, _) -> reportType }
+            .toSet()
+}