+++ delta-coverage-core/src/main/kotlin/io/github/surpsg/deltacoverage/report/jacoco/csv/TextualReportOutputStream.kt
@@ -1,5 +1,6 @@
 package io.github.surpsg.deltacoverage.report.jacoco.csv
 
+import io.github.surpsg.deltacoverage.report.ReportBound
 import io.github.surpsg.deltacoverage.report.ReportType
 import io.github.surpsg.deltacoverage.report.textual.TextualReportFacade
 import java.io.ByteArrayOutputStream
@@ -7,6 +8,7 @@ import java.io.OutputStream
 
 internal class TextualReportOutputStream(
     private val reportType: ReportType,
+    private val reportBound: ReportBound,
     private val outputStream: OutputStream,
 ) : OutputStream() {
 
@@ -17,15 +19,16 @@ internal class TextualReportOutputStream(
     override fun write(b: ByteArray, off: Int, len: Int) = super.write(b, off, len)
 
     override fun close() {
-        outputStream.use { outStream ->
-            val buildContext = TextualReportFacade.BuildContext(
-                coverageDataProvider = CsvSourceRawCoverageDataProvider(byteArrayOutputStream.toByteArray()),
-                reportType = reportType,
-                outputStream = outStream,
-            )
+        outputStream.use { os ->
+            val buildContext = TextualReportFacade.BuildContext {
+                coverageDataProvider = CsvSourceRawCoverageDataProvider(byteArrayOutputStream.toByteArray())
+                reportType = this@TextualReportOutputStream.reportType
+                reportBound = this@TextualReportOutputStream.reportBound
+                outputStream = os
+                shrinkLongClassName = (reportType == ReportType.CONSOLE)
+            }
             TextualReportFacade.generateReport(buildContext)
         }
         super.close()
     }
 }
-