+++ delta-coverage-core/src/main/kotlin/io/github/surpsg/deltacoverage/report/textual/TextualReportFacade.kt
@@ -1,5 +1,7 @@
 package io.github.surpsg.deltacoverage.report.textual
 
+import io.github.surpsg.deltacoverage.lib.kotlin.applyIf
+import io.github.surpsg.deltacoverage.report.ReportBound
 import io.github.surpsg.deltacoverage.report.ReportType
 import io.github.surpsg.deltacoverage.report.textual.TextualReportRenderer.Context
 import java.io.OutputStream
@@ -10,7 +12,6 @@ internal object TextualReportFacade {
     private const val SHRINK_PLACEHOLDER = "..."
 
     private const val PERCENT_MULTIPLIER = 100
-    private const val DELTA_COVERAGE_TITLE = "Delta Coverage Stats"
 
     private const val NA_VALUE = ""
 
@@ -28,25 +29,31 @@ internal object TextualReportFacade {
             rawCoverageData
                 .sortedByDescending { it.linesRatio }
                 .asSequence()
-                .map { it.toValuesCollection() }
+                .map { it.toValuesCollection(buildContext) }
                 .toList()
 
         TextualReportRendererFactory.getBy(buildContext.reportType).render(
             Context {
                 output = buildContext.outputStream
-                title = DELTA_COVERAGE_TITLE
+                title = buildTitle(buildContext)
                 headers = HEADERS
-                footer = rawCoverageData.computeTotal().toValuesCollection()
+                footer = rawCoverageData.computeTotal().toValuesCollection(buildContext)
                 rows = coverageDataValues
             }
         )
     }
 
-    private fun RawCoverageData.toValuesCollection(): List<String> = HEADERS.map { header ->
+    private fun RawCoverageData.toValuesCollection(
+        buildContext: BuildContext,
+    ): List<String> = HEADERS.map { header ->
         when (header) {
             SOURCE_H -> source
-            CLASS_H -> aClass.shrinkClassName(MAX_CLASS_COLUMN_LENGTH)
+
             LINES_H -> linesRatio.formatToPercentage()
+
+            CLASS_H -> aClass.applyIf(buildContext.shrinkLongClassName) {
+                shrinkClassName(MAX_CLASS_COLUMN_LENGTH)
+            }
 
             BRANCHES_H -> {
                 if (branchesTotal == 0) {
@@ -85,10 +92,17 @@ internal object TextualReportFacade {
         }
     }
 
-    internal class BuildContext(
+    private fun buildTitle(buildContext: BuildContext): String = when (buildContext.reportBound) {
+        ReportBound.DELTA_REPORT -> "Delta Coverage Stats"
+        ReportBound.FULL_REPORT -> "Total Coverage Stats"
+    }
+
+    internal class BuildContext private constructor(
         val reportType: ReportType,
+        val reportBound: ReportBound,
         val coverageDataProvider: RawCoverageDataProvider,
         val outputStream: OutputStream,
+        val shrinkLongClassName: Boolean,
     ) {
 
         init {
@@ -98,6 +112,26 @@ internal object TextualReportFacade {
             ) {
                 "Supports only $supportedTypes types"
             }
+        }
+
+        class Builder {
+            lateinit var reportType: ReportType
+            lateinit var reportBound: ReportBound
+            lateinit var coverageDataProvider: RawCoverageDataProvider
+            lateinit var outputStream: OutputStream
+            var shrinkLongClassName: Boolean = false
+
+            fun build() = BuildContext(
+                reportType,
+                reportBound,
+                coverageDataProvider,
+                outputStream,
+                shrinkLongClassName
+            )
+        }
+
+        companion object {
+            operator fun invoke(block: Builder.() -> Unit): BuildContext = Builder().apply(block).build()
         }
     }
 }